refactor: generalize oci-client [NR-515717] (#2156)
