+++ agent-control/tests/on_host/oci_registry.rs
@@ -3,6 +3,7 @@ use crate::on_host::tools::oci_package_manager::TestDataHelper;
 use httpmock::{MockServer, When};
 use newrelic_agent_control::agent_control::run::on_host::OCI_TEST_REGISTRY_URL;
 use newrelic_agent_control::http::config::ProxyConfig;
+use newrelic_agent_control::oci;
 use newrelic_agent_control::package::oci::artifact_definitions::PackageMediaType;
 use newrelic_agent_control::package::oci::downloader::{OCIAgentDownloader, OCIArtifactDownloader};
 use oci_client::client::{ClientConfig, ClientProtocol};
@@ -39,15 +40,15 @@ fn test_download_artifact_from_local_registry_with_oci_registry() {
             .unwrap(),
     );
 
-    let downloader = OCIArtifactDownloader::try_new(
-        ProxyConfig::default(),
-        runtime,
+    let client = oci::Client::try_new(
         ClientConfig {
             protocol: ClientProtocol::Http,
             ..Default::default()
         },
+        ProxyConfig::default(),
     )
     .unwrap();
+    let downloader = OCIArtifactDownloader::new(client, runtime);
 
     let _ = downloader
         .download(&reference, local_agent_data_dir)
@@ -113,16 +114,16 @@ fn test_download_artifact_from_local_registry_using_proxy_with_retries_with_oci_
             .unwrap(),
     );
 
-    let downloader = OCIArtifactDownloader::try_new(
-        proxy_config,
-        runtime,
+    let client = oci::Client::try_new(
         ClientConfig {
             protocol: ClientProtocol::Http,
             ..Default::default()
         },
+        proxy_config,
     )
-    .unwrap()
-    .with_retries(4, Duration::from_millis(100));
+    .unwrap();
+    let downloader =
+        OCIArtifactDownloader::new(client, runtime).with_retries(4, Duration::from_millis(100));
 
     let result = downloader.download(&reference, local_agent_data_dir);
     assert!(result.is_ok());
